requirements.txt
absl-py==0.7.1
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
atari-py==0.2.6
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.6.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.12
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.16.1
gym==0.14.0
gym-gridworlds==0.0.2
gym-unity==0.4.3
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
imageio==2.5.0
importlib-metadata==0.19
incremental==17.5.0
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lockfile==0.12.2
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.6
mkl-random==1.0.2
mkl-service==2.0.2
mlagents-envs==0.9.0
more-itertools==7.2.0
numpy==1.16.4
olefile==0.46
paho-mqtt==1.4.0
Pillow==5.4.1
pluggy==0.12.0
protobuf==3.8.0
py==1.8.0
pyasn1==0.4.6
pyasn1-modules==0.2.1
pycparser==2.19
pycrypto==2.6.1
pyglet==1.3.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyqtgraph==0.10.0
PySocks==1.7.0
pytest==3.10.1
python-dateutil==2.8.0
pytz>=2019.1
scipy==1.3.1
service-identity==18.1.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.2.0
torchsummary==1.5.1
torchvision==0.4.0a0+6b959ee
Twisted==17.9.0
tzlocal==2.0.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
zipp==0.5.2
zope.interface==4.4.3
